recode, code cleanup & rename
